refactor: change `enclave wizard` to `enclave config-set`
